+++ engines/scumm/dimuse_v2/dimuse_v2_cmds.cpp
@@ -24,7 +24,9 @@
 
 namespace Scumm {
 
-int DiMUSE_v2::cmdsHandleCmd(int cmd, int arg_0, int arg_1, int arg_2, int arg_3, int arg_4,
+// We have some intptr_t arguments as commands 28, 29 and 30 actually require pointer arguments
+// Unfortunately this makes function calls for other command a little less pretty...
+int DiMUSE_v2::cmdsHandleCmd(int cmd, int arg_0, intptr_t arg_1, intptr_t arg_2, intptr_t arg_3, intptr_t arg_4,
 	int arg_5, int arg_6, int arg_7, int arg_8, int arg_9,
 	int arg_10, int arg_11, int arg_12, int arg_13) {
 
@@ -57,10 +59,10 @@ int DiMUSE_v2::cmdsHandleCmd(int cmd, int arg_0, int arg_1, int arg_2, int arg_3
 	case 6: // cmds_restore
 		break;
 	case 7:
-		_groupsHandler->setGroupVol(arg_0, arg_1);
+		_groupsHandler->setGroupVol(arg_0, (int)arg_1);
 		break;
 	case 8:
-		cmdsStartSound(arg_0, arg_1);
+		cmdsStartSound(arg_0, (int)arg_1);
 		break;
 	case 9:
 		cmdsStopSound(arg_0);
@@ -71,24 +73,24 @@ int DiMUSE_v2::cmdsHandleCmd(int cmd, int arg_0, int arg_1, int arg_2, int arg_3
 	case 11:
 		return cmdsGetNextSound(arg_0);
 	case 12:
-		cmdsSetParam(arg_0, arg_1, arg_2);
+		cmdsSetParam(arg_0, (int)arg_1, (int)arg_2);
 		break;
 	case 13:
-		return cmdsGetParam(arg_0, arg_1);
+		return cmdsGetParam(arg_0, (int)arg_1);
 	case 14:
-		return _fadesHandler->fadeParam(arg_0, arg_1, arg_2, arg_3, arg_4);
+		return _fadesHandler->fadeParam(arg_0, (int)arg_1, (int)arg_2, (int)arg_3, (int)arg_4);
 	case 15:
-		return cmdsSetHook(arg_0, arg_1);
+		return cmdsSetHook(arg_0, (int)arg_1);
 	case 16:
 		return cmdsGetHook(arg_0);
 	case 17:		
-		return _triggersHandler->setTrigger(arg_0, marker, arg_2, arg_3, arg_4, arg_5, arg_6, arg_7, arg_8, arg_9, arg_10, arg_11, arg_12, arg_13);
+		return _triggersHandler->setTrigger(arg_0, marker, (int)arg_2, (int)arg_3, (int)arg_4, arg_5, arg_6, arg_7, arg_8, arg_9, arg_10, arg_11, arg_12, arg_13);
 	case 18:
-		return _triggersHandler->checkTrigger(arg_0, marker, arg_2);
+		return _triggersHandler->checkTrigger(arg_0, marker, (int)arg_2);
 	case 19:
-		return _triggersHandler->clearTrigger(arg_0, marker, arg_2);
+		return _triggersHandler->clearTrigger(arg_0, marker, (int)arg_2);
 	case 20:
-		return _triggersHandler->deferCommand(arg_0, arg_1, arg_2, arg_3, arg_4, arg_5, arg_6, arg_7, arg_8, arg_9, arg_10, arg_11, arg_12, arg_13);
+		return _triggersHandler->deferCommand(arg_0, (int)arg_1, (int)arg_2, (int)arg_3, (int)arg_4, arg_5, arg_6, arg_7, arg_8, arg_9, arg_10, arg_11, arg_12, arg_13);
 	case 21:
 	case 22:
 	case 23:
@@ -96,15 +98,15 @@ int DiMUSE_v2::cmdsHandleCmd(int cmd, int arg_0, int arg_1, int arg_2, int arg_3
 		// Empty opcodes
 		return 0;
 	case 25:
-		return waveStartStream(arg_0, arg_1, arg_2);
+		return waveStartStream(arg_0, (int)arg_1, (int)arg_2);
 	case 26:
-		return waveSwitchStream(arg_0, arg_1, arg_2, arg_3, arg_4);
+		return waveSwitchStream(arg_0, (int)arg_1, (int)arg_2, (int)arg_3, (int)arg_4);
 	case 27:
 		return waveProcessStreams();
 	case 28:
 		return waveQueryStream(arg_0, (int *)arg_1, (int *)arg_2, (int *)arg_3, (int *)arg_4);
 	case 29:
-		return waveFeedStream(arg_0, (int *)arg_1, arg_2, arg_3);
+		return waveFeedStream(arg_0, (int *)arg_1, (int)arg_2, (int)arg_3);
 	case 30:
 		return waveLipSync(arg_0, arg_1, arg_2, (int32 *)arg_3, (int32 *)arg_4);
 	default:
